requirements.txt
openai
openai-whisper
gTTS
ipython
google-colab
torch
